Add support for running instant commands quickly.
Removing backgroundScript and adding scriptsQueue. Each command
in the commandQueue gets to run for one turn.  If they do not
finish in one turn then they are added to the scriptsQueue. Commands
in the scripts finish in FIFO order.  This allows for instant commands
that finish quickly and long running commands that take time. Stop
is an instant command.